Inline matches module into dockerfile
